+++ opentelemetry-instrumentation-tower/examples/hyper-http-service/src/main.rs
@@ -61,10 +61,8 @@ async fn main() {
         .build();
 
     global::set_meter_provider(meter_provider);
-    // init our otel metrics middleware
-    let global_meter = global::meter(SERVICE_NAME);
+    
     let otel_metrics_service_layer = otel_tower_metrics::HTTPMetricsLayerBuilder::builder()
-        .with_meter(global_meter)
         .build()
         .unwrap();